Refactor comment.dart and post_detail_screen.dart
This commit refactors the comment.dart file to update the vote and parent properties to accept nullable integer values. It also refactors the post_detail_screen.dart file to import necessary dependencies and add a BlocBuilder widget to display comments based on the state of the DetailPostBloc. These changes improve the code structure and enhance the functionality of the app by allowing users to view and interact with comments on posts.